V2.0_0324

+++ src/Jason/GoalsInfo.java
@@ -12,11 +12,11 @@ public class GoalsInfo extends ParseJason{
         this.updateUrl = "https://studev.groept.be/api/a21ib2d02/GoalsSet/";
     }
 
-    public int getGoal() {
+    public double getGoal() {
         JSONArray array = new JSONArray(makeGETRequest(getUrl));
         JSONObject curObject = array.getJSONObject(0);
 
-        return Integer.parseInt(curObject.getString("day_goal"));
+        return Double.parseDouble(curObject.getString("day_goal"));
     }
 
     public int getTimer() {
@@ -26,7 +26,7 @@ public class GoalsInfo extends ParseJason{
         return Integer.parseInt(curObject.getString("timer"));
     }
 
-    public void updateGoal(int goal, int timer){
+    public void updateGoal(double goal, int timer){
         makeGETRequest(updateUrl + goal + "/" + timer);
     }
 }
